+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/contracts/domain/model/ContractType.java
@@ -1,0 +1,5 @@
+package org.eclipse.tractusx.traceability.contracts.domain.model;
+
+public enum ContractType {
+    ASSET_AS_PLANNED, ASSET_AS_BUILT, NOTIFICATION;
+}