+++ CHANGELOG.md
@@ -15,6 +15,7 @@ _**For better traceability add the corresponding GitHub issue number in each cha
 - #753 Refactored message history in notification detail view
 - XXX updated local deployment documentation
 - #1037 extended autocomplete api by contractAgreementId
+- #985 Added function to save Contracts based on notification contractAgreementIds into the database
 
 ### Added
 - #832 added policymanagement list view, creator and editor